Renamed huart2 huart in Adafruit file, made gpsData fixed point only

+++ MITCH_Common/System/System_Inc/interface_structs.h
@@ -27,8 +27,8 @@ typedef struct gpsData {
 	char nmeaGGA[MAX_NMEA];
 	char nmeaRMC[MAX_NMEA];
 	uint8_t fix;
-	float alt;
-	float speed;
+	int32_t alt;
+	int32_t speed;
 	int utc;
 	bool hasUpdate;
 	bool lock;
